Customized Prefix
This PR adds the functionality and interface
to allow for alternative S3 prefixes to be
used for the remote storage repository.
Previously, a fixed `pipper/` prefix was used,
which has now become the default when a value
is not applied explicitly from the command line
or from a saved repository.